ZHA: Add Material Icons and add in Connexion page for email and password
