+++ with-docker-compose-app/next-app/package.json
@@ -7,10 +7,14 @@
   },
   "dependencies": {
     "18": "^0.0.0",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
     "@fortawesome/fontawesome-svg-core": "^6.6.0",
     "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@lucia-auth/adapter-sqlite": "^3.0.2",
+    "@mui/icons-material": "^6.1.5",
+    "@mui/material": "^6.1.5",
     "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
     "better-sqlite3": "^11.3.0",
